I think that GitHub repo is ready.
Guys, congrats, we were hacking almost all of 24 hours and it was a lot of fun.
Even thought, there are still a few small bugs and we could add more integrations, we can call this a demo.
Personally, I had a lot of fun and I am very excited for cuHacking next year, when, if you agree, we can all be in the same team ;) :) .
Also, I am glad, that we agreed to continue with this project, after hackathon as well and I can not wait to see it what will it turn out.
That is all from me for now.
Thank you for awesome hackathon,
Best Reguards,

Srdjan.